Inventory shifting RE:#1

+++ routes/rfid_conn.js
@@ -73,48 +73,52 @@ router.post('/', (req, res) => {
                     return res.status(500).json({ message: 'Failed to insert data', error: err });
                 }
 
-                const { prodID: iProdID, start: iStart, end: iEnd } = plisReader(result[0]["PLIS"]);
-                const { prodID, start, end } = plisReader(jPlis);
+                if (result.length !== 0) {
+                    const { prodID: iProdID, start: iStart, end: iEnd } = plisReader(result[0]["PLIS"]);
+                    const { prodID, start, end } = plisReader(jPlis);
 
-                const rfidQuery = 'INSERT INTO rfid_logs (RFID, WID, PLIS, TSTMP, LOC) VALUES (?, ?, ?, NOW(), ?);';
-                const transferQuery = 'INSERT INTO inventory (WID, PPID, PLIS, CNT) VALUES (?, ?, ?, ?);';
+                    const rfidQuery = 'INSERT INTO rfid_logs (RFID, WID, PLIS, TSTMP, LOC) VALUES (?, ?, ?, NOW(), ?);';
+                    const transferQuery = 'INSERT INTO inventory (WID, PPID, PLIS, CNT) VALUES (?, ?, ?, ?);';
 
-                if (prodID === iProdID && (start === iStart && end <= iEnd)) {
-                    const rWritePlis = plisWriter(prodID, start, end);
-                    if (end !== iEnd) {
-                        const writePlis = plisWriter(prodID, end + 1, iEnd);
-                        const writeCnt = iEnd - end;
+                    if (prodID === iProdID && (start === iStart && end <= iEnd)) {
+                        const rWritePlis = plisWriter(prodID, start, end);
+                        if (end !== iEnd) {
+                            const writePlis = plisWriter(prodID, end + 1, iEnd);
+                            const writeCnt = iEnd - end;
 
-                        const updateQuery = 'UPDATE inventory SET PLIS = ?, CNT = ?;';
-                        connection.query(updateQuery, [writePlis, writeCnt], (err, reses) => {
-                            connection.query(rfidQuery, [rfid, twid, rWritePlis, loc], (err, reses1) => {
-                                connection.query(transferQuery, [twid, iProdID, rWritePlis, writeCnt], (err, reses2) => {
-                                    connection.close();
-                                    if (err) {
-                                        return res.status(500).json({ message: 'Failed to insert data', error: err });
-                                    }
+                            const updateQuery = 'UPDATE inventory SET PLIS = ?, CNT = ?;';
+                            connection.query(updateQuery, [writePlis, writeCnt], (err, reses) => {
+                                connection.query(rfidQuery, [rfid, twid, rWritePlis, loc], (err, reses1) => {
+                                    connection.query(transferQuery, [twid, iProdID, rWritePlis, writeCnt], (err, reses2) => {
+                                        connection.close();
+                                        if (err) {
+                                            return res.status(500).json({ message: 'Failed to insert data', error: err });
+                                        }
 
-                                    res.status(200).json({ message: "Part of the inventory has been transferred successfully" });
+                                        res.status(200).json({ message: "Part of the inventory has been transferred successfully" });
+                                    });
                                 });
                             });
-                        });
+                        } else {
+                            const deleteQuery = 'DELETE FROM inventory WHERE EID = ?;';
+                            connection.query(deleteQuery, [result[0]["EID"]], (err, reses) => {
+                                connection.query(rfidQuery, [rfid, twid, rWritePlis, loc], (err, reses1) => {
+                                    connection.query(transferQuery, [twid, iProdID, rWritePlis, writeCnt], (err, reses2) => {
+                                        connection.close();
+                                        if (err) {
+                                            return res.status(500).json({ message: 'Failed to insert data', error: err });
+                                        }
+
+                                        res.status(200).json({ message: "All products transferred successfully" });
+                                    });
+                                });
+                            });
+                        }
                     } else {
-                        const deleteQuery = 'DELETE FROM inventory WHERE EID = ?;';
-                        connection.query(deleteQuery, [result[0]["EID"]], (err, reses) => {
-                            connection.query(rfidQuery, [rfid, twid, rWritePlis, loc], (err, reses1) => {
-                                connection.query(transferQuery, [twid, iProdID, rWritePlis, writeCnt], (err, reses2) => {
-                                    connection.close();
-                                    if (err) {
-                                        return res.status(500).json({ message: 'Failed to insert data', error: err });
-                                    }
-
-                                    res.status(200).json({ message: "All products transferred successfully" });
-                                });
-                            });
-                        });
+                        res.status(200).json({ message: "Product Range Out of Index" });
                     }
                 } else {
-                    res.status(200).json({ message: "Product Range Out of Index" });
+                    res.status(200).json({ message: "Inventory Empty" });
                 }
             });
         }
